SVY-19261 Invalid scoping of const variable
https://github.com/mozilla/rhino/issues/326

introduced a feature flag: FEATURE_LEGACY_CONST_MODE

to get the legacy behavior (default is the modern behavior)

so that const are threated as a let but then in readonly mode

for this the parser will create also for const the symbol table that is
then later picked up to add them to the scope.

The IRFactory transforBlock() will place the current function in
requiresActivation because if you are in a block and there are lets of
const we really need this behavior
Else the const and even the let are not scoped and are accessible
outside of the block, except if you have a inner function because that
would trigger the requiredActivation also.

NodeTransformer now adds a special IndexObject (instead of just a name)
that also hold the type, so that the ScriptRuntime when filling the
object literal (the scope) knows if it needs to define a constant or
just a let

Also ScriptableObject put will check also the new feature flag if it
should throw whena const is reassigned (instead of silently ignore it)